requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
croniter==5.0.1
discord-py-interactions==5.13.2
discord-typings==0.9.0
emoji==2.14.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tomli==2.1.0
typing_extensions==4.12.2
yarl==1.17.2
